+++ CHANGELOG.md
@@ -19,6 +19,7 @@
  - PR #123 Remove driver install from ci scripts
  - PR #131 Use YYMMDD tag in nightly build
  - PR #137 Replace CFFI python bindings with Cython
+ - PR #127 Use Memory Resource classes for allocations
 
 ## Bug Fixes
 
@@ -26,6 +27,7 @@
  - PR #110 Fix Skip Test Functionality
  - PR #125 Fixed order of private variables in LogIt
  - PR #139 Expose `_make_finalizer` python API needed by cuDF
+ - PR #142 Fix ignored exceptions in Cython
 
 
 # RMM 0.9.0 (21 August 2019)
@@ -34,7 +36,7 @@
 
  - PR #96 Added `device_memory_resource` for beginning of overhaul of RMM design
  - PR #103 Add and use unified build script
- 
+
 ## Improvements
 
  - PR #111 Streamline CUDA_REL environment variable